refactor(patch-detail): extract title and descr and form in a new cmp
