+++ src/webviews/src/components/PatchDetail.vue
@@ -1,13 +1,11 @@
 <script setup lang="ts">
 import {
   provideVSCodeDesignSystem,
-  vsCodeButton,
-  vsCodeTextArea,
   vsCodePanels,
   vsCodePanelTab,
   vsCodePanelView,
 } from '@vscode/webview-ui-toolkit'
-import { ref, computed, watchEffect, nextTick } from 'vue'
+import { ref, computed } from 'vue'
 import {
   breakpointsTailwind,
   useBreakpoints,
@@ -23,21 +21,13 @@ import { scrollToTemplateRef } from '@/utils/scrollToTemplateRef'
 import PatchStatusBadge from '@/components/PatchStatusBadge.vue'
 import PatchMajorEvents from '@/components/PatchMajorEvents.vue'
 import PatchDetailButtons from '@/components/PatchDetailButtons.vue'
-import PatchMetadata from '@/components/PatchMetadata.vue'
+import PatchDetailTitleDescription from '@/components/PatchDetailTitleDescription.vue'
 import PatchDetailActivity from '@/components/PatchDetailActivity.vue'
 import PatchDetailRevision from '@/components/PatchDetailRevision.vue'
-import Markdown from '@/components/Markdown.vue'
 
-provideVSCodeDesignSystem().register(
-  vsCodeButton(),
-  vsCodeTextArea(),
-  vsCodePanels(),
-  vsCodePanelTab(),
-  vsCodePanelView(),
-)
+provideVSCodeDesignSystem().register(vsCodePanels(), vsCodePanelTab(), vsCodePanelView())
 
-const { patch, authors, firstRevision, latestRevision, patchEditForm } =
-  storeToRefs(usePatchDetailStore())
+const { patch, authors, firstRevision, latestRevision } = storeToRefs(usePatchDetailStore())
 
 const revisionTabRef = ref<HTMLElement>()
 const revisionSectionRef = ref<HTMLElement>()
@@ -92,61 +82,6 @@ function assembleRevisionOptionLabel(revision: Revision): string {
 }
 
 // TODO: show "edited" indicators + timestamp (on hover) or full-blown list of edits, for each revision, comment, etc anything that has edits
-
-const formEl = ref<HTMLElement>()
-const titleTextAreaEl = ref<HTMLElement>()
-const descrTextAreaEl = ref<HTMLElement>()
-
-// TODO: maninak add button to toggle between markdown preview and editing
-// TODO: maninak add ctrl/cmd + (B | I | E | K) listeners to toggle bold | italics | backtick-quote | link
-watchEffect(() => {
-  const titleEl = titleTextAreaEl.value?.shadowRoot?.querySelector('textarea')
-  const descrEl = descrTextAreaEl.value?.shadowRoot?.querySelector('textarea')
-  const els = [titleEl, descrEl].filter(Boolean)
-
-  els.forEach((el) => {
-    useEventListener(
-      el,
-      'keydown',
-      (ev) => {
-        if (ev.key === 'Escape') {
-          patchEditForm.value.isEditing = false
-        }
-        if (ev.key === 'Enter' && (ev.ctrlKey || ev.metaKey)) {
-          submitPatchEditForm()
-        }
-      },
-      { passive: true },
-    )
-    useEventListener(el, 'focus', () => resetTextAreaHeight(el), { passive: true })
-    useEventListener(el, 'input', () => resetTextAreaHeight(el), { passive: true })
-  })
-})
-
-function resetTextAreaHeight(el: HTMLTextAreaElement) {
-  el.style.height = ''
-  el.style.height = `${el.scrollHeight + 3}px` // additional offset needed to avoid scrollbar
-  formEl.value?.scrollIntoView({ block: 'end', behavior: 'instant' })
-}
-
-function beginPatchEditing() {
-  patchEditForm.value.title = patchEditForm.value.title || patch.value.title
-  patchEditForm.value.descr = patchEditForm.value.descr || patch.value.revisions[0].description
-  patchEditForm.value.isEditing = true
-  nextTick(() => titleTextAreaEl.value?.focus())
-}
-
-function submitPatchEditForm() {
-  patchEditForm.value.isEditing = false
-  patch.value.title = patchEditForm.value.title.trim() // TODO: maninak delete and notify extension to change title instead
-  patchEditForm.value.title = ''
-  patch.value.revisions[0].description = patchEditForm.value.descr.trim() // TODO: maninak delete and notify extension to change title instead
-  patchEditForm.value.descr = ''
-}
-
-function discardPatchEditForm() {
-  patchEditForm.value.isEditing = false
-}
 </script>
 
 <template>
@@ -165,72 +100,7 @@ function discardPatchEditForm() {
     <main
       class="grid grid-rows-subgrid grid-cols-subgrid row-span-3 sm:row-span-2 sm:col-span-2"
     >
-      <section class="flex flex-col gap-y-4" style="grid-area: section-patch">
-        <PatchMetadata />
-        <div v-if="!patchEditForm.isEditing" class="max-w-fit flex flex-col gap-y-4 group">
-          <div class="flex gap-x-2">
-            <h1 class="my-0 text-3xl font-mono"><Markdown :source="patch.title" /></h1>
-            <vscode-button
-              appearance="icon"
-              title="Edit Patch Title and Description"
-              class="opacity-0 focus:opacity-100 group-hover:opacity-100"
-              @click="beginPatchEditing"
-            >
-              <!-- eslint-disable-next-line vue/no-deprecated-slot-attribute -->
-              <span class="codicon codicon-edit"></span>
-            </vscode-button>
-          </div>
-          <Markdown :source="firstRevision.description" class="text-sm" />
-        </div>
-        <form
-          v-else
-          @submit.prevent
-          ref="formEl"
-          name="Edit patch title and description"
-          class="pb-2 w-fit flex flex-col gap-y-3"
-        >
-          <vscode-text-area
-            ref="titleTextAreaEl"
-            :value="patchEditForm.title"
-            @input="(ev: CustomEvent) => (patchEditForm.title = ev.target?._value)"
-            name="patch title"
-            rows="1"
-            cols="100"
-            resize="vertical"
-            maxlength="400"
-          >
-            Patch Title:
-          </vscode-text-area>
-          <vscode-text-area
-            ref="descrTextAreaEl"
-            :value="patchEditForm.descr"
-            @input="(ev: CustomEvent) => (patchEditForm.descr = ev.target?._value)"
-            name="patch description"
-            rows="5"
-            cols="100"
-            resize="vertical"
-            maxlength="500000"
-          >
-            Patch Description:
-          </vscode-text-area>
-          <div class="w-full flex flex-row-reverse justify-start gap-x-2">
-            <vscode-button
-              appearance="primary"
-              title="Save Changes to Radicle"
-              @click="submitPatchEditForm"
-            >
-              Save
-            </vscode-button>
-            <vscode-button
-              appearance="secondary"
-              title="Stop Editing and Discard Current Changes"
-              @click="discardPatchEditForm"
-            >
-              Discard
-            </vscode-button>
-          </div>
-        </form>
-      </section>
+      <PatchDetailTitleDescription />
 
       <vscode-panels
         v-if="isWindowNarrowerThanSm"
@@ -292,17 +162,6 @@ function discardPatchEditForm() {
 </template>
 
 <style scoped>
-vscode-text-area::part(control) {
-  min-height: 5ch;
-  max-height: min(80ch, 65vh);
-  field-sizing: content;
-  @apply font-mono text-sm;
-}
-
-vscode-text-area::part(label) {
-  margin-bottom: 0.5em;
-}
-
 .grid-areas-patch {
   grid-template-areas:
     'header           '
